requirements.txt
numpy
pymongo
pandas
nba_api
SQLAlchemy
pg8000
mlflow
pyaml
polars
tqdm